refactor: Move FrontController and RequestHandler to normal package
